Check for null values from getRunningAppProcesses()
It can return null if there are no running app processes.

It is not entirely clear to me how one can end up having no
running app processes when the check for ProcessPhoenix happens from
within an app, but I've seen multiple crashes now from it, so it
must be possible.